+++ lib/src/dkd/factory.dart
@@ -34,7 +34,7 @@ import 'proc.dart';
 
 
 // -----------------------------------------------------------------------------
-// GeneralContentProcessorFactory (Concrete CPU Factory)
+//  GeneralContentProcessorFactory (Concrete CPU Factory)
 // -----------------------------------------------------------------------------
 
 /// General implementation of [ContentProcessorFactory] with caching support.