+++ lib/src/msg/instant_delegate.dart
@@ -30,7 +30,7 @@
  */
 import 'dart:typed_data';
 
-import 'package:dimp/dimp.dart';
+import 'package:dimp/protocol.dart';
 
 import '../crypto/bundle.dart';