+++ lib/src/msg/instant_packer.dart
@@ -30,7 +30,9 @@
  */
 import 'dart:typed_data';
 
-import 'package:dimp/dimp.dart';
+import 'package:dimp/crypto.dart';
+import 'package:dimp/protocol.dart';
+import 'package:dimp/dkd.dart';
 
 import '../crypto/bundle.dart';
 import 'instant_delegate.dart';