+++ lib/src/msg/secure_packer.dart
@@ -30,7 +30,8 @@
  */
 import 'dart:typed_data';
 
-import 'package:dimp/dimp.dart';
+import 'package:dimp/crypto.dart';
+import 'package:dimp/protocol.dart';
 
 import '../crypto/bundle.dart';
 import 'secure_delegate.dart';